chore: Split up results classes and definitions
